Create Readme.md
**Pi-star local Pocsag Gateway**

Pi-star is a wonderful initiative for radio amateurs. The link with the DAPnet certainly also has added value.

Unfortunately, it also happens that you are in a location where the internet is poor or not available.

To still be able to send Pocsag messages locally, I made a simple script in PHP. You can place this script on the pi-star web server on your Raspberry Pi.